move DockingDataset and work on plotting functions
