+++ covid_moonshot_ml/docking/__init__.py
@@ -1,2 +1,2 @@
-from . import mcs
-from . import docking
+#from . import mcs
+#from . import docking
