+++ scripts/analyze_docking_results.py
@@ -10,7 +10,7 @@ from covid_moonshot_ml.datasets.utils import load_openeye_pdb, \
     get_ligand_rmsd_openeye, get_ligand_RMSD_mdtraj, load_openeye_sdf
 from covid_moonshot_ml.schema import ExperimentalCompoundDataUpdate, \
     EnantiomerPairList
-from covid_moonshot_ml.docking.schema import DockingDataset
+from covid_moonshot_ml.docking.analysis import DockingDataset
 
 
 def get_args():
